fix(api): promote MFA state only while its version is still pending
The version check sat in front of an unconditional write, so it promised
more than it enforced: StartMFAEnrollment or DisableMFA could commit in
between, and the confirmation would still write what it had read —
promoting an enrollment nobody was waiting for, or reviving the factor an
administrator had just cleared for a locked-out user.

The predicate now rides on the write. UpdateUserMFAConfigIfPending
updates the row only while its stored pending version still matches, and
a confirmation that no longer matches is refused rather than committed.
Clearing the pending slot on promotion also makes a replay — a
double-submit, a tab that never learned it succeeded — find nothing to
promote instead of rewriting the factor.

EnableMFA keeps its read-side check: it writes nothing, and the promotion
re-checks against the row it updates.

Also folds the six methods into user_service.go rather than a separate
file, keeping one file per service.

Test: TestUpdateUserMFAConfigIfPending covers a stale version, the
matching one, and a replay; it fails without the predicate on the write.
